remove old lib
