--- Meshkraft/src/main/java/com/artlabs/meshkraft/data/model/Assets.kt
@@ -1,21 +1,0 @@
-package com.artlabs.meshkraft.data.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-/**
- * @author Mert Gölcü
- * @date 12.11.2021
- */
-
-@Serializable
-data class Assets(
-    @SerialName("glb")
-    val glb: Glb?
-)
-
-@Serializable
-data class Glb(
-    val url: String,
-    val size: Double
-)
