+++ Meshkraft/src/main/java/com/artlabs/meshkraft/data/model/Product.kt
@@ -1,58 +1,49 @@
-package com.mertgolcu.meshkraft.AR
+package com.artlabs.meshkraft.data.model
 
-import com.artlabs.meshkraft.data.model.Assets
-import com.artlabs.meshkraft.data.model.Config
-import com.artlabs.meshkraft.data.model.Image
-import com.artlabs.meshkraft.data.model.Model
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 /**
  * @author Mert Gölcü
- * @date 10.09.2021
+ * @date 21.01.2022
  */
+
 @Serializable
 data class Product(
-    @SerialName("category")
-    val category: String?,
-    @SerialName("config")
-    val config: Config?,
-    @SerialName("createdAt")
-    val createdAt: String?,
-    @SerialName("_id")
     val _id: String?,
     @SerialName("id")
-    val id: String?,
-    @SerialName("images")
-    val images: List<Image>?,
-    @SerialName("models")
     val models: List<Model>?,
     @SerialName("name")
     val name: String?,
-    @SerialName("preview")
-    val preview: String?,
-    @SerialName("published_at")
-    val publishedAt: String?,
     @SerialName("SKU")
     val sKU: String?,
-    @SerialName("updatedAt")
-    val updatedAt: String?,
-    @SerialName("__v")
-    val v: Int?,
-    @SerialName("assets")
-    val assets:Assets?
 )
 
+@Serializable
+data class Model(
+    @SerialName("file")
+    val `file`: File?,
+    @SerialName("_id")
+    val size: String?,
+    @SerialName("status")
+    val status: String?,
+    @SerialName("type")
+    val type: String?,
+)
 
-
-
-
-
-
-
-
-
-
-
-
-
+@Serializable
+data class File(
+    @SerialName("ext")
+    val ext: String?,
+    @SerialName("hash")
+    val hash: String?,
+    @SerialName("height")
+    val height: Any?,
+    @SerialName("size")
+    val size: Double?,
+    @SerialName("url")
+    val url: String?,
+    @SerialName("width")
+    val width: Any?
+)
