+++ app/src/main/java/com/mertgolcu/testapp/MainActivity.kt
@@ -1,6 +1,8 @@
 package com.mertgolcu.testapp
 
 import android.os.Bundle
+import android.widget.Button
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.artlabs.meshkraft.IMeshkraftState
 import com.artlabs.meshkraft.Meshkraft
@@ -8,33 +10,42 @@ import com.artlabs.meshkraft.data.model.Mode
 
 //KOTLIN
 class MainActivity : AppCompatActivity() {
+
+    private lateinit var textView: TextView
+    private lateinit var button: Button
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        Meshkraft.setApiKey("API_KEY")
+        textView = findViewById(R.id.text_view)
+        button = findViewById(R.id.button)
 
-        Meshkraft.startArSession(
-            context = this,
-            sku = "TST00078SKU41",
-            mode = Mode.AR_ONLY,
-            listener = object : IMeshkraftState {
-                override fun onLoading() {
-                    //loading state
+        Meshkraft.setApiKey("phTzrnocCq2bFfWAbtz8CFxuCAXvfVyaK3VrCeJpUD")
+
+        button.setOnClickListener {
+            Meshkraft.startArSession(
+                context = this,
+                sku = "HBV00000MML21",
+                mode = Mode.PREFERRED_3D,
+                listener = object : IMeshkraftState {
+                    override fun onLoading() {
+                        textView.text = "Yükleniyor..."
+                    }
+
+                    override fun onFail(message: String) {
+                        textView.text = "Hata oluştu : $message"
+                    }
+
+                    override fun onFinish() {
+                        textView.text = "Tamamlandı"
+                    }
+
                 }
+            )
+        }
 
-                override fun onFail(message: String) {
-                    // any error or fail
-                }
 
-                override fun onFinish() {
-                    // on finish all
-                }
-
-            }
-        )
-
-        Meshkraft.startBasicArSession(this, "TST00078SKU41")
+        // Meshkraft.startBasicArSession(this, "TST00078SKU41")
 
 
     }
